fix(analysis): support binary structures and switching legs
Preserve fractional AiiDA kind weights as chemical composition and obtain formulas from StructureData rather than applying integer gcd arithmetic. Parse Calphy leg1 and leg2 switching files, retain the leg identity, and integrate against the complete reference energy when it is represented by multiple dU_ref terms. Retrieved analyses now expose stable AiiDA repository identities instead of temporary extraction paths.

+++ src/aiida_melting/analysis/records.py
@@ -14,7 +14,7 @@ class ResultRecord:
     process_uuid: str
     ctime: str
     formula: str | None
-    composition: dict[str, int]
+    composition: dict[str, float]
     elements: tuple[str, ...]
     pressure_gpa: float | None
     method: str | None
